requirements.txt
anyio==4.8.0
idna==3.10
PyYAML==6.0.2
sniffio==1.3.1
typing_extensions==4.12.2
watchfiles==1.0.4
